[config] refactoring
