+++ config/src/main/scala/net/kemitix/thorp/config/ConfigValidationException.scala
@@ -1,5 +1,5 @@
 package net.kemitix.thorp.config
 
 final case class ConfigValidationException(
-    errors: List[ConfigValidation]
+    errors: Seq[ConfigValidation]
 ) extends Exception
