+++ config/src/test/scala/net/kemitix/thorp/config/ConfigOptionTest.scala
@@ -1,6 +1,6 @@
 package net.kemitix.thorp.config
 
-import net.kemitix.thorp.domain.Sources
+import net.kemitix.thorp.domain.{Sources, TemporaryFolder}
 import net.kemitix.thorp.filesystem.FileSystem
 import org.scalatest.FunSpec
 import zio.DefaultRuntime
